Refactor utility methods from EdgeFactory to TimeseriesUtility
